Fix: move `INITIATE_DOWNLOAD_TASK` action to Dashboard
