+++ webapp/app/containers/App/constants.ts
@@ -56,9 +56,6 @@ export const LOAD_DOWNLOAD_LIST_FAILURE = 'davinci/Download/LOAD_DOWNLOAD_LIST_F
 export const DOWNLOAD_FILE = 'davinci/Download/DOWNLOAD_FILE'
 export const DOWNLOAD_FILE_SUCCESS = 'davinci/Download/DOWNLOAD_FILE_SUCCESS'
 export const DOWNLOAD_FILE_FAILURE = 'davinci/Download/DOWNLOAD_FILE_FAILURE'
-export const INITIATE_DOWNLOAD_TASK = 'davinci/Download/INITIATE_DOWNLOAD_TASK'
-export const INITIATE_DOWNLOAD_TASK_SUCCESS = 'davinci/Download/INITIATE_DOWNLOAD_TASK_SUCCESS'
-export const INITIATE_DOWNLOAD_TASK_FAILURE = 'davinci/Download/INITIATE_DOWNLOAD_TASK_FAILURE'
 
 export const CREATE_ORGANIZATION_PROJECT = 'davinci/permission/CREATE_ORGANIZATION_PROJECT'
 export const DELETE_ORGANIZATION_PROJECT = 'davinci/permission/DELETE_ORGANIZATION_PROJECT'
