+++ webapp/app/containers/Dashboard/constants.ts
@@ -50,6 +50,10 @@ enum Types {
   LOAD_DASHBOARD_ITEM_DATA_FAILURE = 'davinci/Dashboard/LOAD_DASHBOARD_ITEM_DATA_FAILURE',
   LOAD_BATCH_DATA_WITH_CONTROL_VALUES = 'davinci/Dashboard/LOAD_BATCH_DATA_WITH_CONTROL_VALUES',
 
+  INITIATE_DOWNLOAD_TASK = 'davinci/Download/INITIATE_DOWNLOAD_TASK',
+  INITIATE_DOWNLOAD_TASK_SUCCESS = 'davinci/Download/INITIATE_DOWNLOAD_TASK_SUCCESS',
+  INITIATE_DOWNLOAD_TASK_FAILURE = 'davinci/Download/INITIATE_DOWNLOAD_TASK_FAILURE',
+
   LOAD_DASHBOARD_SHARE_LINK = 'davinci/Dashboard/LOAD_DASHBOARD_SHARE_LINK',
   LOAD_DASHBOARD_SHARE_LINK_SUCCESS = 'davinci/Dashboard/LOAD_DASHBOARD_SHARE_LINK_SUCCESS',
   LOAD_DASHBOARD_SHARE_LINK_FAILURE = 'davinci/Dashboard/LOAD_DASHBOARD_SHARE_LINK_FAILURE',
